move/JS
moved second version of code to new .js and .html file as I've added new datasets that are linked to kecodes to support keyboard:

https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/keyCode

added the ability to do multiple calculations as once with multiple operators

Shows calc history at the top

removed CE button as it is not needed for the assignment.

Only thing that is still missing on my code is the backspace button.